requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
da-vinci==0.3.0
Django==4.0.1
django-filter==21.1
django-rest-swagger==2.2.0
django-thumbnails==0.7.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
shortuuid==1.0.8
simplejson==3.17.6
sorl-thumbnail==12.3
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
